+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/MetroFirCheckers.kt
@@ -7,6 +7,7 @@ import dev.zacsweers.metro.compiler.fir.checkers.AsContributionChecker
 import dev.zacsweers.metro.compiler.fir.checkers.AssistedInjectChecker
 import dev.zacsweers.metro.compiler.fir.checkers.BindingContainerCallableChecker
 import dev.zacsweers.metro.compiler.fir.checkers.BindingContainerClassChecker
+import dev.zacsweers.metro.compiler.fir.checkers.ConflictingAnnotationRolesChecker
 import dev.zacsweers.metro.compiler.fir.checkers.CreateGraphChecker
 import dev.zacsweers.metro.compiler.fir.checkers.DefaultBindingChecker
 import dev.zacsweers.metro.compiler.fir.checkers.DependencyGraphChecker
@@ -53,12 +54,11 @@ internal class MetroFirCheckers(session: FirSession) : FirAdditionalCheckersExte
     object : ExpressionCheckers() {
       override val annotationCheckers: Set<FirAnnotationChecker>
         get() {
-          return if (
-            session.metroFirBuiltIns.options.interopAnnotationsNamedArgSeverity.isEnabled
-          ) {
-            setOf(InteropAnnotationChecker)
-          } else {
-            super.annotationCheckers
+          return buildSet {
+            add(ConflictingAnnotationRolesChecker)
+            if (session.metroFirBuiltIns.options.interopAnnotationsNamedArgSeverity.isEnabled) {
+              add(InteropAnnotationChecker)
+            }
           }
         }